Gemini  App to Retrival SQL Data/requirements.txt
streamlit
google-generstiveai
python-dotenv